Added agent switching
Added a way to switch between agents by releasing the spacebar. An indicator will light up for as long as the spacebar is held. When released a raycast is done from the current agent to check for others (the raycast ignores all obstacles). Next, I will add a little highlight to show what agent is currently being controlled, and then a way to move groups of agents.